requirements.txt
numpy>=1.26.4
pandas>=2.1.0
scikit-learn>=1.4.0
shap>=0.44.0
requests>=2.31.0
python-dotenv>=1.0.1
matplotlib>=3.8.0
seaborn>=0.13.0
joblib>=1.3.2

streamlit>=1.30.0
folium>=0.16.0
streamlit-folium>=0.20.0
branca>=0.7.0
plotly>=5.18.0
